Add a simple initial IMGUI library (imi.lua) to create the plugin UI
We have two possible paths to create the plugin UI:

1) Fully extend Aseprite Lua API to access/layout all kind of widgets
   just like in C++ code (too complex to extend the API at this moment
   and too complex to create the extension then)
2) Create a library to handle widgets inside a Dialog:canvas()

Here we have a chance to create an IMGUI library to simplify the
creation and maintenance of the plugin code. We don't need too many
widgets at the moment (mainly viewports to see tiles/attachments with
scrollbars) and toggleable buttons to see sub folders/categories.