Fix crash undoing the update of the sprite DB
Regression introduced in 3350c82884b2d008902a6c8afdf602d1bb9653e8, we
cannot update properties when we navigate the undo. We could filter
the property change using "if not ev.fromUndo then..." but in this
case we prefer to avoid modifying the Base Set if the tile is
missing (we should provide other way to add/fix missing tiles in the
Base Set, which shouldn't be a common scenario).