Fix new category not added to properties.categories
'table.insert(properties.categories, id)' do not add the desired 'id'.
It was discovered that it cannot added an element via table.insert()
when the table argument is the 'properties' of the element (layers,
tiles, etc.).

This can be reproduced when a new tilemap layer is added, it's default
tileset wasn't added as a category.