Fix error creating new layer when the active layer is a tilemap layer moved from the origin
As NewLayer was using the current site grid bounds to create the
tileset, if we have a moved tilemap from the origin the tilemap grid
will be used a the grid to create a new layer (which is not
desirable).

A new gridBounds parameter was added to NewLayer to fix this.