Remove Edit Anchors and Edit Attachments options
The new Insert Joint command replaces the Edit Anchors functionality.
We'll add any extra functionality dynamically in the main dialog
instead of showing another temporal dialog.

The Edit Attachments command was removed because it wasn't used. We
could bring it back if it's necessary, but we'll try to add all new UI
on the active canvas/editor. Any command that helps to edit things
that are not visible (e.g. a to edit attachments/anchors that are not
used in the sprite) at the very end will not be useful (because what
it really matters is the visible animation/attachments).