When we change the active attachment, align only children of that attachment (#94)
The Align Anchors commands still can align all attachments in the
current frame from the top-most (free/without parent) attachments.
But now when we double-click an attachment to put it on the canvas,
only its children will be aligned (and the attachment is pivoted from
the ref point if it exists in the old and new attachment).