Attachments categories synchronization loss '2' (fix #138)
Fixed category desync due to automatic attachment adding in empty cel
when the activeTileset is not a base category.

Root-cause:
Aseprite core adds an automatic tile to the active tileset when the cel
is empty and we draw on the canvas. So the Sprite_afteraddtile() event
only updates the current tileset. This action culminates in a de-sync
with the other tileset categories.